chore(vir): Lock verus-lib dependency version
We want to have a commit with an explicit version lock of the verus-lib
dependency. This is being done because we will be working on updating
the verus-lib dependency which will perhaps result in major changes for
this project.

Therefore we want to have a special revision which is stable and uses a
stable version of verus-lib.